float ram

+++ modulos/cpu_grupo18/cpu_grupo18.c
@@ -22,7 +22,7 @@ static int cpu_g18_main(struct seq_file *m, void *v)
     size_t process_counter = 0;
     for_each_process(procesos)
     {
-        unsigned long ram = 0;
+        float ram = 0;
         if (procesos->mm && procesos->mm->total_vm)
         {
             ram = procesos->mm->total_vm;
